requirements.txt
google-genai>=0.5.0
streamlit>=1.32.0
langchain-chroma>=0.1.0
chromadb>=0.4.22
langchain>=0.1.0
langchain-google-genai>=0.0.11
langchain-community>=0.0.29
langchain-text-splitters>=0.0.1
python-dotenv>=1.0.0
rich>=13.7.0
pytest>=8.0.0
